Move PreferenceDSL
# Conflicts:
#	app/src/main/java/eu/kanade/tachiyomi/ui/extension/SettingsExtensionsController.kt
#	app/src/main/java/eu/kanade/tachiyomi/ui/setting/SettingsAboutController.kt
